feat: add custom panic hook
This replaces the default panic hook by one that will run the default
hook and send the panic occurrence to PostHog, with privacy sensitive
information erased (e.g., database names and user directories). The
hook also asks whether the user would like to send their name and e-mail
in case it is deemed necessary to enter in contact to obtain further
information.

The hook is only enabled when telemetry is enabled.