feat: cache GitHub GET requests
This caches the GitHub GET requests in the Synth configuration file, as
the new "version_check_delay" field. When running a Synth command, if the
field is present and has not expired, no additional GET request is
made. If the field is not present or _has_ expired, the request is
made as usual and the version check delay date is saved in the
configuration file.

This should considerably improve the responsiveness of Synth when
running short duration commands.