fix: Fix panic on 'Nan' in csv_str_to_value
This fixes the panic that was thrown when csv_str_to_value attempted to
convert an 'invalid' float value into a JSON number. If the float value
was not representable as a JSON number, it would panic.
The new behaviour is to fall back to representing 'invalid' float values
as Strings, but still attempt to parse them as numbers first.