fix: Fix panic during CSV import when encountering invalid floats (#433)
* fix: Fix panic on 'Nan' in csv_str_to_value

This fixes the panic that was thrown when csv_str_to_value attempted to
convert an 'invalid' float value into a JSON number. If the float value
was not representable as a JSON number, it would panic.
The new behaviour is to fall back to representing 'invalid' float values
as Strings, but still attempt to parse them as numbers first.

* chore: Update 'log' dep version to 0.4.18

Updating the version of the 'log' dependency to 0.4.18 to fix
compilation issues on recent Nightly toolchains.  See
https://github.com/async-rs/async-std/issues/1058 for a similar issue.

* chore: Run rustfmt, add clippy allow

Ran rustfmt, there were some existing changes that ended up getting
reformatted.
Added an explicit `#[allow(clippy::default_constructed_unit_structs)]`
in `core/src/graph/number.rs` - `OsRng::default()` is used in a unit
test.